% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental_saves_R.R
\name{annotate_ws}
\alias{annotate_ws}
\title{Annotate Work Space Reference Table}
\usage{
annotate_ws(
  subset,
  annotation = NULL,
  write = FALSE,
  file = "ws_table.ref",
  rds.folder = "rds"
)
}
\arguments{
\item{subset}{A logical expression to select specific objects}

\item{annotation}{Character string with the annotation}

\item{write}{If TRUE, annotations are actually written in the file. If FALSE, a filtered data.frame with the proposed annotations is returned.}

\item{file}{Name of the file in the folder containing the information about the saved workspaces}

\item{rds.folder}{Name of the folder where .rds files are stored}
}
\value{
The file is updated with the new annotated table.
}
\description{
Annotate lines in the Work Space Reference Table
}
