% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental_saves_R.R
\name{load_incremental}
\alias{load_incremental}
\title{Load incremental}
\usage{
load_incremental(subset, metadata.file = "ws_table.ref")
}
\arguments{
\item{subset}{A logical expression to select specific objects}

\item{metadata.file}{Name of the file in the folder containing the information about the saved workspaces}
}
\value{
A loaded workspace in the Gloval Environment
}
\description{
Load, potentially selecting specific objects, all the incremental saves in the project folder
}
