% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental_saves_R.R
\name{purge_ws_table}
\alias{purge_ws_table}
\title{Purge Work Space Reference Table}
\usage{
purge_ws_table(subset, file = "ws_table.ref", remove = FALSE)
}
\arguments{
\item{subset}{A logical expression to select specific objects}

\item{file}{Name of the file in the folder containing the information about the saved workspaces}

\item{remove}{If TRUE, objects are actually removed. If FALSE, a list of objects that would be removed is returned.}
}
\value{
.RData files are updated removing the selected objects and the Work Space Reference Table is updated.
}
\description{
Remove objects from the saved .RData files and update the Work Space Reference Table
}
