% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental_saves_R.R
\name{save_incremental}
\alias{save_incremental}
\title{Save incremental}
\usage{
save_incremental(file, metadata.file = "ws_table.ref")
}
\arguments{
\item{file}{Name of the .RData file to store the saved workspace}

\item{metadata.file}{Name of the file in the folder containing the information about the saved workspaces}
}
\value{
An incremental save of the workspace is stored in a file on the project's folder
}
\description{
Save only the objects from the workspace that are not already stored in other .RData files in the project folder
}
